pyproject.toml
[tool.poetry]
name = "schema-manager"
version = "0.1.1"
description = "Schema manager tool for enterprise governance"
authors = ["branquif <branquif@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
lark-parser = "^0.6.6"
jinja2 = "^2.10"
mimesis = "^3.1"

[tool.poetry.dev-dependencies]
pytest = "^4.2"
ipython = "^7.2"
tabview = "^1.4"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
isort = "^4.3"
flake8 = "^3.7"
autoflake = "^1.2"
mypy = "^0.670.0"
pydocstyle = "^3.0"
rope = "^0.12.0"
pyflakes = "^2.1"
pynvim = "^0.3.2"
invoke = "^1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
